refactor(view, constant) : MenuOption 위치 이동
